pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agno-docs-mcp"
version = "0.1.0"
description = "MCP server for accessing Agno framework documentation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Uzair", email = "uzair@example.com" }
]
keywords = ["mcp", "agno", "documentation", "llm", "ai-agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
agno-docs-mcp = "agno_docs_mcp.server:main"

[project.urls]
Homepage = "https://github.com/agno-agi/agno-docs-mcp"
Documentation = "https://docs.agno.com"
Repository = "https://github.com/agno-agi/agno-docs-mcp"

[tool.hatch.build.targets.wheel]
packages = ["src/agno_docs_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/.docs",
    "/README.md",
]
